skillomate-AI/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai>=1.0.0
PyAudio==0.2.11
SpeechRecognition==3.10.0
gTTS==2.3.2
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
matplotlib==3.7.2
seaborn==0.12.2
numpy==1.24.3
Pillow==10.0.0